+++ src/LossFunctions/CategoricalCrossEntropy.cpp
@@ -2,7 +2,7 @@
 
 namespace neural_net {
 
-double CategoricalCrossEntropy::Loss(const Matrix& present, const Matrix& expected) const {
+float CategoricalCrossEntropy::Loss(const Matrix& present, const Matrix& expected) const {
     assert(present.cols() == expected.cols() && present.rows() == expected.rows());
     return -(expected.array() * (present.array() + kEpsilon).log()).sum() / present.rows();
 }
